+++ packages/placeholder-pdfkit010/src/pdfkitAddPlaceholder.js
@@ -21,6 +21,7 @@ import {PDFKitReferenceMock} from './pdfkitReferenceMock';
 * @property {string} [byteRangePlaceholder]
 * @property {string} [subFilter] One of SUBFILTER_* from \@signpdf/utils
 * @property {number[]} [widgetRect] [x1, y1, x2, y2] widget rectangle
+* @property {string} [appName] Name of the application generating the signature
 */
 
 /**
@@ -49,6 +50,7 @@ export const pdfkitAddPlaceholder = ({
     byteRangePlaceholder = DEFAULT_BYTE_RANGE_PLACEHOLDER,
     subFilter = SUBFILTER_ADOBE_PKCS7_DETACHED,
     widgetRect = [0, 0, 0, 0],
+    appName = undefined,
 }) => {
     /* eslint-disable no-underscore-dangle,no-param-reassign */
     // Generate the signature placeholder
@@ -68,6 +70,12 @@ export const pdfkitAddPlaceholder = ({
         ContactInfo: new String(contactInfo), // eslint-disable-line no-new-wrappers
         Name: new String(name), // eslint-disable-line no-new-wrappers
         Location: new String(location), // eslint-disable-line no-new-wrappers
+        Prop_Build: {
+            Filter: {Name: 'Adobe.PPKLite'},
+            ...(
+                appName ? {App: {Name: appName}} : {}
+            ),
+        },
     });
 
     // Check if pdf already contains acroform field